Platform/Sgi: add helper macros for ITS, SMMUv3 and DMA IORT nodes
Add preprocessor macros for ITS, SMMUv3 and DMA named component nodes
of IORT table that can be used to describe the IO topology connected
to the IO Virtualization block. An IO virtualization block could be
used to connect PCIe root bus or non-PCIe devices.

For a non-PCIe IO-Virtualization block, there are DMA masters
connected to SMMUv3 in the following manner -

  0x0  -------------   0x30000   ------------
      |  DMA device |---------->|    SMMU    |
      |    (ID)     |           | (StreamID) |
  0x9  -------------   0x30009   ------------

The DMA master consists of 8 data channels and 1 instruction channel.
The IDs emitted by DMA go to the SMMUv3 as StreamID and the SMMUv3
programs the corresponding translation contexts.

The PCIe root bus or non-PCIe devices are connected to one of the
x16/x8/x4_1/x4_0 ports of IO virtualization block. Each of these
ports have a base DeviceID that is added to the RequesterID or
StreamID of the devices connected to ports to create the IDs sent
to the SMMUv3 and ITS.

Stream ID coming at SMMUv3 is calculated as below:
  Stream ID = DMA Channel Index + Base PCI port index, e.g.

For example, for channel 1 of DMA0 device (connected to x16 port
whose baseID is 0x30000, the streamID that SMMUv3 sees is:
  (1 + 0x30000) = 0x30001.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>